dogbowl_api/requirements.txt
aniso8601==9.0.1
click==8.0.0
Flask==2.0.0
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
importlib-metadata==4.0.1
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
marshmallow==3.12.1
marshmallow-sqlalchemy==0.25.0
numpy==1.20.3
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2021.1
RPi.GPIO==0.7.1a4
scipy==1.6.3
six==1.16.0
SQLAlchemy==1.4.15
typing-extensions==3.10.0.0
Werkzeug==2.0.0
zipp==3.4.1
